test fixture again

+++ tests/data/test_datareader.py
@@ -38,6 +38,9 @@ def setup_data():
 
 
 def teardown(filename=''):
+    print('===========================filename', filename, type(filename))
+    if type(filename) != str:
+        return
     if os.path.isfile(filename):
         os.remove(filename)
 
@@ -49,18 +52,16 @@ def h5file():
     print('>>>>>>>>>>>>>>>>>>>>>>>>>>>>filename', filename)
     data, target = setup_data()
 
-    hf = h5py.File(filename, 'w')
+    with h5py.File(filename, 'w') as hf:
+        for i in range(10):
+            start, end = i * 100, (i + 1) * 100
+            group = hf.create_group('fold_{0}'.format(i))
+            group.create_dataset('input', data=data[start: end])
+            group.create_dataset('target', data=target[start: end])
 
-    for i in range(10):
-        start, end = i * 100, (i + 1) * 100
-        group = hf.create_group('fold_{0}'.format(i))
-        group.create_dataset('input', data=data[start: end])
-        group.create_dataset('target', data=target[start: end])
-
-    hf.close()
     print('<<<<<<<<<<<<<<<<<<>>>>>>>filename', filename)
     yield copy(filename)
-    print('<<<<<<<<<<<<<<<<<<=======filename', filename)
+    print('<<<<<<<<<<<<<<<<<<=======filename', filename, type(filename))
     teardown(filename)
